axum_extra::extract

Struct WithRejection

pub struct WithRejection<E, R>(pub E, pub PhantomData<R>);

Extractor for customizing extractor rejections

WithRejection wraps another extractor and gives you the result. If the extraction fails, the Rejection is transformed into R and returned as a response

E is expected to implement FromRequest

R is expected to implement IntoResponse and From<E::Rejection>

Example

use axum::extract::rejection::JsonRejection;
use axum::response::{Response, IntoResponse};
use axum::Json;
use axum_extra::extract::WithRejection;
use serde::Deserialize;

struct MyRejection { /* ... */ }

impl From<JsonRejection> for MyRejection {
    fn from(rejection: JsonRejection) -> MyRejection {
        // ...
    }
}

impl IntoResponse for MyRejection {
    fn into_response(self) -> Response {
        // ...
    }
}
#[derive(Debug, Deserialize)]
struct Person { /* ... */ }

async fn handler(
    // If the `Json` extractor ever fails, `MyRejection` will be sent to the
    // client using the `IntoResponse` impl
    WithRejection(Json(Person), _): WithRejection<Json<Person>, MyRejection>
) { /* ... */ }

Tuple Fields
0: E1: PhantomData<R>
Implementations

impl<E, R> WithRejection<E, R>

pub fn into_inner(self) -> E

Returns the wrapped extractor