Crate axum_extra

Extra utilities for axum.

Feature flags

axum-extra uses a set of [feature flags] to reduce the amount of compiled and optional dependencies.

The following optional features are available:

Name	Description	Default?

async-read-body	Enables the AsyncReadBody body	No

attachment	Enables the Attachment response	No

cookie	Enables the CookieJar extractor	No

cookie-private	Enables the PrivateCookieJar extractor	No

cookie-signed	Enables the SignedCookieJar extractor	No

cookie-key-expansion	Enables the Key::derive_from method	No

erased-json	Enables the ErasedJson response	No

error-response	Enables the InternalServerError response	No

form	Enables the Form extractor	No

json-deserializer	Enables the JsonDeserializer extractor	No

json-lines	Enables the JsonLines extractor and response	No

multipart	Enables the Multipart extractor	No

protobuf	Enables the Protobuf extractor and response	No

query	Enables the Query extractor	No

tracing	Log rejections from built-in extractors	Yes

typed-routing	Enables the TypedPath routing utilities	No

typed-header	Enables the [TypedHeader] extractor and response	No

file-stream	Enables the FileStream response	No

Modules

body

Additional bodies.

either

Either* types for combining extractors or responses into a single type.

extract

Additional extractors.

handler

Additional handler utilities.

middleware

Additional middleware utilities.

response

Additional types for generating responses.

routing

Additional types for defining routes.